% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday problem probability}
\usage{
birthday(x)
}
\arguments{
\item{x}{The number of people.}
}
\value{
The probability of least two people sharing a birthday.
}
\description{
Calculates the probability for the birthday problem.
}
\examples{
birthday(23)
}
