% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylab5.R
\name{mybin}
\alias{mybin}
\title{mylab5}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{# How many times to iterate}

\item{n}{# Number of values}

\item{p}{# Probability}
}
\value{
# A table and plot
}
\description{
mylab5
}
\examples{
mylab5(100, 15, 0.5)
}
