% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations of the pulled sample}

\item{x}{the dataset sampled from}

\item{fun}{the function being used}

\item{alpha}{the confidence interval tail lengths}

\item{cx}{controls the size of the text in the barplot}

\item{...}{}
}
\value{
a barplot of the confidence interval given
}
\description{
myboot2
}
\examples{
myboot2(10000,x=sam,fun="mean",alpha=0.05,xlab="mean",col="Purple",cx=1.5)
}
