% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{mycltp}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations of the sample size}

\item{lambda}{the margin of success}

\item{...}{extra parameters}
}
\value{
multiple graphs of the probability of myclt
}
\description{
mycltp
}
\examples{
mycltp(n=2,iter=10000, lambda=4)
}
