% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Average Value}

\item{sigma}{Standard Deviation}

\item{a}{Interval Upper Limit}
}
\value{
A list including entered values and probability, as well as a graph
}
\description{
myncurve
}
\examples{
myncurve(mu = 1, sigma = 2, a = 3)
}
