% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{ntickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The amount of seats on the plane}

\item{gamma}{The maximum allowed probability of overbooking the plane}

\item{p}{The probability that the person who bought a ticket shows up}
}
\value{
A graph of the discrete binomial distribution, a graph of the normal
approximation and a list of nd, nc, N and gamma
}
\description{
ntickets
}
\examples{
ntickets(N=400,gamma = 0.02, p = 0.95)
}
